+++ OUDS/Core/Components/Sources/_/Extensions/View+Typography.swift
@@ -24,7 +24,7 @@ extension View {
     /// - Parameter theme: The current `OUDSTheme` to use to load the current font family and the suitable font semantic token.
     /// - Returns some View: The current `View` but with new typography applied
     public func displayLarge(_ theme: OUDSTheme) -> some View {
-        modifier(TypographyModifier(family: theme.fontFamily, font: theme.fonts.displayLarge))
+        modifier(TypographyModifier(family: theme.fonts.familyDisplay, font: theme.fonts.displayLarge))
     }
 
     /// Modifies the current `View` to apply a *display medium* typography.
@@ -32,7 +32,7 @@ extension View {
     /// - Parameter theme: The current `OUDSTheme` to use to load the current font family and the suitable font semantic token.
     /// - Returns some View: The current `View` but with new typography applied
     public func displayMedium(_ theme: OUDSTheme) -> some View {
-        modifier(TypographyModifier(family: theme.fontFamily, font: theme.fonts.displayMedium))
+        modifier(TypographyModifier(family: theme.fonts.familyDisplay, font: theme.fonts.displayMedium))
     }
 
     /// Modifies the current `View` to apply a *display small* typography.
@@ -40,7 +40,7 @@ extension View {
     /// - Parameter theme: The current `OUDSTheme` to use to load the current font family and the suitable font semantic token.
     /// - Returns some View: The current `View` but with new typography applied
     public func displaySmall(_ theme: OUDSTheme) -> some View {
-        modifier(TypographyModifier(family: theme.fontFamily, font: theme.fonts.displaySmall))
+        modifier(TypographyModifier(family: theme.fonts.familyDisplay, font: theme.fonts.displaySmall))
     }
 
     /// Modifies the current `View` to apply a *heading x large* typography.
@@ -48,7 +48,7 @@ extension View {
     /// - Parameter theme: The current `OUDSTheme` to use to load the current font family and the suitable font semantic token.
     /// - Returns some View: The current `View` but with new typography applied
     public func headingXLarge(_ theme: OUDSTheme) -> some View {
-        modifier(TypographyModifier(family: theme.fontFamily, font: theme.fonts.headingXLarge))
+        modifier(TypographyModifier(family: theme.fonts.familyHeading, font: theme.fonts.headingXLarge))
     }
 
     /// Modifies the current `View` to apply a *heading large* typography.
@@ -56,7 +56,7 @@ extension View {
     /// - Parameter theme: The current `OUDSTheme` to use to load the current font family and the suitable font semantic token.
     /// - Returns some View: The current `View` but with new typography applied
     public func headingLarge(_ theme: OUDSTheme) -> some View {
-        modifier(TypographyModifier(family: theme.fontFamily, font: theme.fonts.headingLarge))
+        modifier(TypographyModifier(family: theme.fonts.familyHeading, font: theme.fonts.headingLarge))
     }
 
     /// Modifies the current `View` to apply a *heading medium* typography.
@@ -64,7 +64,7 @@ extension View {
     /// - Parameter theme: The current `OUDSTheme` to use to load the current font family and the suitable font semantic token.
     /// - Returns some View: The current `View` but with new typography applied
     public func headingMedium(_ theme: OUDSTheme) -> some View {
-        modifier(TypographyModifier(family: theme.fontFamily, font: theme.fonts.headingMedium))
+        modifier(TypographyModifier(family: theme.fonts.familyHeading, font: theme.fonts.headingMedium))
     }
 
     /// Modifies the current `View` to apply a *heading small* typography.
@@ -72,7 +72,7 @@ extension View {
     /// - Parameter theme: The current `OUDSTheme` to use to load the current font family and the suitable font semantic token.
     /// - Returns some View: The current `View` but with new typography applied
     public func headingSmall(_ theme: OUDSTheme) -> some View {
-        modifier(TypographyModifier(family: theme.fontFamily, font: theme.fonts.headingSmall))
+        modifier(TypographyModifier(family: theme.fonts.familyHeading, font: theme.fonts.headingSmall))
     }
 
     /// Modifies the current `View` to apply a *body default large* typography.
@@ -80,7 +80,7 @@ extension View {
     /// - Parameter theme: The current `OUDSTheme` to use to load the current font family and the suitable font semantic token.
     /// - Returns some View: The current `View` but with new typography applied
     public func bodyDefaultLarge(_ theme: OUDSTheme) -> some View {
-        modifier(TypographyModifier(family: theme.fontFamily, font: theme.fonts.bodyDefaultLarge))
+        modifier(TypographyModifier(family: theme.fonts.familyBody, font: theme.fonts.bodyDefaultLarge))
     }
 
     /// Modifies the current `View` to apply a *body default medium* typography.
@@ -88,7 +88,7 @@ extension View {
     /// - Parameter theme: The current `OUDSTheme` to use to load the current font family and the suitable font semantic token.
     /// - Returns some View: The current `View` but with new typography applied
     public func bodyDefaultMedium(_ theme: OUDSTheme) -> some View {
-        modifier(TypographyModifier(family: theme.fontFamily, font: theme.fonts.bodyDefaultMedium))
+        modifier(TypographyModifier(family: theme.fonts.familyBody, font: theme.fonts.bodyDefaultMedium))
     }
 
     /// Modifies the current `View` to apply a *body default small* typography.
@@ -96,7 +96,7 @@ extension View {
     /// - Parameter theme: The current `OUDSTheme` to use to load the current font family and the suitable font semantic token.
     /// - Returns some View: The current `View` but with new typography applied
     public func bodyDefaultSmall(_ theme: OUDSTheme) -> some View {
-        modifier(TypographyModifier(family: theme.fontFamily, font: theme.fonts.bodyDefaultSmall))
+        modifier(TypographyModifier(family: theme.fonts.familyBody, font: theme.fonts.bodyDefaultSmall))
     }
 
     /// Modifies the current `View` to apply a *body moderate large* typography.
@@ -104,7 +104,7 @@ extension View {
     /// - Parameter theme: The current `OUDSTheme` to use to load the current font family and the suitable font semantic token.
     /// - Returns some View: The current `View` but with new typography applied
     public func bodyModerateLarge(_ theme: OUDSTheme) -> some View {
-        modifier(TypographyModifier(family: theme.fontFamily, font: theme.fonts.bodyModerateLarge))
+        modifier(TypographyModifier(family: theme.fonts.familyBody, font: theme.fonts.bodyModerateLarge))
     }
 
     /// Modifies the current `View` to apply a *body moderate medium* typography.
@@ -112,7 +112,7 @@ extension View {
     /// - Parameter theme: The current `OUDSTheme` to use to load the current font family and the suitable font semantic token.
     /// - Returns some View: The current `View` but with new typography applied
     public func bodyModerateMedium(_ theme: OUDSTheme) -> some View {
-        modifier(TypographyModifier(family: theme.fontFamily, font: theme.fonts.bodyModerateMedium))
+        modifier(TypographyModifier(family: theme.fonts.familyBody, font: theme.fonts.bodyModerateMedium))
     }
 
     /// Modifies the current `View` to apply a *body moderate small* typography.
@@ -120,7 +120,7 @@ extension View {
     /// - Parameter theme: The current `OUDSTheme` to use to load the current font family and the suitable font semantic token.
     /// - Returns some View: The current `View` but with new typography applied
     public func bodyModerateSmall(_ theme: OUDSTheme) -> some View {
-        modifier(TypographyModifier(family: theme.fontFamily, font: theme.fonts.bodyModerateSmall))
+        modifier(TypographyModifier(family: theme.fonts.familyBody, font: theme.fonts.bodyModerateSmall))
     }
 
     /// Modifies the current `View` to apply a *body strong large* typography.
@@ -128,7 +128,7 @@ extension View {
     /// - Parameter theme: The current `OUDSTheme` to use to load the current font family and the suitable font semantic token.
     /// - Returns some View: The current `View` but with new typography applied
     public func bodyStrongLarge(_ theme: OUDSTheme) -> some View {
-        modifier(TypographyModifier(family: theme.fontFamily, font: theme.fonts.bodyStrongLarge))
+        modifier(TypographyModifier(family: theme.fonts.familyBody, font: theme.fonts.bodyStrongLarge))
     }
 
     /// Modifies the current `View` to apply a *body strong medium* typography.
@@ -136,7 +136,7 @@ extension View {
     /// - Parameter theme: The current `OUDSTheme` to use to load the current font family and the suitable font semantic token.
     /// - Returns some View: The current `View` but with new typography applied
     public func bodyStrongMedium(_ theme: OUDSTheme) -> some View {
-        modifier(TypographyModifier(family: theme.fontFamily, font: theme.fonts.bodyStrongMedium))
+        modifier(TypographyModifier(family: theme.fonts.familyBody, font: theme.fonts.bodyStrongMedium))
     }
 
     /// Modifies the current `View` to apply a *body strong small* typography.
@@ -144,7 +144,7 @@ extension View {
     /// - Parameter theme: The current `OUDSTheme` to use to load the current font family and the suitable font semantic token.
     /// - Returns some View: The current `View` but with new typography applied
     public func bodyStrongSmall(_ theme: OUDSTheme) -> some View {
-        modifier(TypographyModifier(family: theme.fontFamily, font: theme.fonts.bodyStrongSmall))
+        modifier(TypographyModifier(family: theme.fonts.familyBody, font: theme.fonts.bodyStrongSmall))
     }
 
     /// Modifies the current `View` to apply a *label default x large* typography.
@@ -152,7 +152,7 @@ extension View {
     /// - Parameter theme: The current `OUDSTheme` to use to load the current font family and the suitable font semantic token.
     /// - Returns some View: The current `View` but with new typography applied
     public func labelDefaultXLarge(_ theme: OUDSTheme) -> some View {
-        modifier(TypographyModifier(family: theme.fontFamily, font: theme.fonts.labelDefaultXLarge))
+        modifier(TypographyModifier(family: theme.fonts.familyLabel, font: theme.fonts.labelDefaultXLarge))
     }
 
     /// Modifies the current `View` to apply a *label default large* typography.
@@ -160,7 +160,7 @@ extension View {
     /// - Parameter theme: The current `OUDSTheme` to use to load the current font family and the suitable font semantic token.
     /// - Returns some View: The current `View` but with new typography applied
     public func labelDefaultLarge(_ theme: OUDSTheme) -> some View {
-        modifier(TypographyModifier(family: theme.fontFamily, font: theme.fonts.labelDefaultLarge))
+        modifier(TypographyModifier(family: theme.fonts.familyLabel, font: theme.fonts.labelDefaultLarge))
     }
 
     /// Modifies the current `View` to apply a *label default medium* typography.
@@ -168,7 +168,7 @@ extension View {
     /// - Parameter theme: The current `OUDSTheme` to use to load the current font family and the suitable font semantic token.
     /// - Returns some View: The current `View` but with new typography applied
     public func labelDefaultMedium(_ theme: OUDSTheme) -> some View {
-        modifier(TypographyModifier(family: theme.fontFamily, font: theme.fonts.labelDefaultMedium))
+        modifier(TypographyModifier(family: theme.fonts.familyLabel, font: theme.fonts.labelDefaultMedium))
     }
 
     /// Modifies the current `View` to apply a *label default small* typography.
@@ -176,7 +176,7 @@ extension View {
     /// - Parameter theme: The current `OUDSTheme` to use to load the current font family and the suitable font semantic token.
     /// - Returns some View: The current `View` but with new typography applied
     public func labelDefaultSmall(_ theme: OUDSTheme) -> some View {
-        modifier(TypographyModifier(family: theme.fontFamily, font: theme.fonts.labelDefaultSmall))
+        modifier(TypographyModifier(family: theme.fonts.familyLabel, font: theme.fonts.labelDefaultSmall))
     }
 
     /// Modifies the current `View` to apply a *label moderate x large* typography.
@@ -184,7 +184,7 @@ extension View {
     /// - Parameter theme: The current `OUDSTheme` to use to load the current font family and the suitable font semantic token.
     /// - Returns some View: The current `View` but with new typography applied
     public func labelModerateXLarge(_ theme: OUDSTheme) -> some View {
-        modifier(TypographyModifier(family: theme.fontFamily, font: theme.fonts.labelModerateXLarge))
+        modifier(TypographyModifier(family: theme.fonts.familyLabel, font: theme.fonts.labelModerateXLarge))
     }
 
     /// Modifies the current `View` to apply a *label moderate large* typography.
@@ -192,7 +192,7 @@ extension View {
     /// - Parameter theme: The current `OUDSTheme` to use to load the current font family and the suitable font semantic token.
     /// - Returns some View: The current `View` but with new typography applied
     public func labelModerateLarge(_ theme: OUDSTheme) -> some View {
-        modifier(TypographyModifier(family: theme.fontFamily, font: theme.fonts.labelModerateLarge))
+        modifier(TypographyModifier(family: theme.fonts.familyLabel, font: theme.fonts.labelModerateLarge))
     }
 
     /// Modifies the current `View` to apply a *label moderate medium* typography.
@@ -200,7 +200,7 @@ extension View {
     /// - Parameter theme: The current `OUDSTheme` to use to load the current font family and the suitable font semantic token.
     /// - Returns some View: The current `View` but with new typography applied
     public func labelModerateMedium(_ theme: OUDSTheme) -> some View {
-        modifier(TypographyModifier(family: theme.fontFamily, font: theme.fonts.labelModerateMedium))
+        modifier(TypographyModifier(family: theme.fonts.familyLabel, font: theme.fonts.labelModerateMedium))
     }
 
     /// Modifies the current `View` to apply a *label moderate small* typography.
@@ -208,7 +208,7 @@ extension View {
     /// - Parameter theme: The current `OUDSTheme` to use to load the current font family and the suitable font semantic token.
     /// - Returns some View: The current `View` but with new typography applied
     public func labelModerateSmall(_ theme: OUDSTheme) -> some View {
-        modifier(TypographyModifier(family: theme.fontFamily, font: theme.fonts.labelModerateSmall))
+        modifier(TypographyModifier(family: theme.fonts.familyLabel, font: theme.fonts.labelModerateSmall))
     }
 
     /// Modifies the current `View` to apply a *label strong x large* typography.
@@ -216,7 +216,7 @@ extension View {
     /// - Parameter theme: The current `OUDSTheme` to use to load the current font family and the suitable font semantic token.
     /// - Returns some View: The current `View` but with new typography applied
     public func labelStrongXLarge(_ theme: OUDSTheme) -> some View {
-        modifier(TypographyModifier(family: theme.fontFamily, font: theme.fonts.labelStrongXLarge))
+        modifier(TypographyModifier(family: theme.fonts.familyLabel, font: theme.fonts.labelStrongXLarge))
     }
 
     /// Modifies the current `View` to apply a *label strong large* typography.
@@ -224,7 +224,7 @@ extension View {
     /// - Parameter theme: The current `OUDSTheme` to use to load the current font family and the suitable font semantic token.
     /// - Returns some View: The current `View` but with new typography applied
     public func labelStrongLarge(_ theme: OUDSTheme) -> some View {
-        modifier(TypographyModifier(family: theme.fontFamily, font: theme.fonts.labelStrongLarge))
+        modifier(TypographyModifier(family: theme.fonts.familyLabel, font: theme.fonts.labelStrongLarge))
     }
 
     /// Modifies the current `View` to apply a *label strong medium* typography.
@@ -232,7 +232,7 @@ extension View {
     /// - Parameter theme: The current `OUDSTheme` to use to load the current font family and the suitable font semantic token.
     /// - Returns some View: The current `View` but with new typography applied
     public func labelStrongMedium(_ theme: OUDSTheme) -> some View {
-        modifier(TypographyModifier(family: theme.fontFamily, font: theme.fonts.labelStrongMedium))
+        modifier(TypographyModifier(family: theme.fonts.familyLabel, font: theme.fonts.labelStrongMedium))
     }
 
     /// Modifies the current `View` to apply a *label strong small* typography.
@@ -240,7 +240,7 @@ extension View {
     /// - Parameter theme: The current `OUDSTheme` to use to load the current font family and the suitable font semantic token.
     /// - Returns some View: The current `View` but with new typography applied
     public func labelStrongSmall(_ theme: OUDSTheme) -> some View {
-        modifier(TypographyModifier(family: theme.fontFamily, font: theme.fonts.labelStrongSmall))
+        modifier(TypographyModifier(family: theme.fonts.familyLabel, font: theme.fonts.labelStrongSmall))
     }
 
     /// Modifies the current `View` to apply a *code medium* typography.
@@ -248,7 +248,7 @@ extension View {
     /// - Parameter theme: The current `OUDSTheme` to use to load the current font family and the suitable font semantic token.
     /// - Returns some View: The current `View` but with new typography applied
     public func codeMedium(_ theme: OUDSTheme) -> some View {
-        modifier(TypographyModifier(family: theme.fontFamily, font: theme.fonts.codeMedium))
+        modifier(TypographyModifier(family: theme.fonts.familyCode, font: theme.fonts.codeMedium))
     }
 
     // MARK: - Font modifier