+++ OUDS/Core/Themes/Sosh/Sources/SoshTheme.swift
@@ -145,7 +145,8 @@ public final class SoshTheme: OUDSTheme, @unchecked Sendable {
                    resourcesBundle: Bundle.SoshTheme,
                    name: Self.name,
                    fontFamily: SoshBrandFontRawTokens.familyDefault,
-                   tuning: Tuning.Sosh)
+                   tuning: Tuning.Sosh,
+                   hasTypographyHeadingLargeMarker: false)
 
         registerFonts()
     }