+++ OUDS/Core/Themes/Wireframe/Sources/WireframeTheme.swift
@@ -141,7 +141,8 @@ public final class WireframeTheme: OUDSTheme, @unchecked Sendable {
                    resourcesBundle: Bundle.WireframeTheme,
                    name: Self.name,
                    fontFamily: WireframeBrandFontRawTokens.familyDefault,
-                   tuning: Tuning.Wireframe)
+                   tuning: Tuning.Wireframe,
+                   hasTypographyHeadingLargeMarker: true)
 
         registerFonts()
     }